ltp: correctly set _root to LTPROOT
When passing in another LTPROOT path than the default one '/opt/ltp/',
kirk complains that it can't find the LTPROOT since it looks for it in
the default LTPROOT path.  By setting the LTPROOT with '--env
LTPROOT=/my/own/path/to/ltp' it will corretly pick up the new path and
run the tests.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>